requirements.txt
flask==3.0.0
python-pptx==0.6.21
openai==0.28.1
python-dotenv==1.0.0
